// src/app/features/pelanggan-request/models/pelanggan-request.dto.ts
export interface PelangganRequestDto {
  id: number;
  id_aoldatabase: number;
  id_aoldb: number;
  id_cabang: number;
  nama_administrator: string;
  email_administrator: string;
  kategori_pelanggan: string;
  nama_pelanggan: string;
  kontak_pelanggan: string;
  kontak_email: string;
  kontak_handphone: string;
  kontak_alamat: string;
  alamat_kirim_pelanggan: string;
  kode_desa: string;
  desa_pelanggan: string;
  kecamatan_pelanggan: string;
  kota_pelanggan: string;
  propinsi_pelanggan: string;
  negara_pelanggan: string;
  latlong_pelanggan: string;
  alamat_tagih_pelanggan: string;
  telp_pelanggan: string;
  email_pelanggan: string;
  handphone_pelanggan: string;
  nik_pelanggan: string;
  npwp_pelanggan: string;
  keterangan_pelanggan: string;
  status_usaha: string;
  luas_tempat: number;
  status_tempat: string;
  image_pelanggan: string;
  image_ktp: string;
  id_userrequest: number;
  potensial_pelanggan: string;
  id_uservalidasi: number;
  validasi_pelanggan: string;
  supplier_pelanggan: string;
  produk_pelanggan: string;
  feedback_lingkungan: string;
  feedback_salesman: string;
  feedback_supplier: string;
  plafon_pelanggan: number;
  id_userrelease: number;
  release_pelanggan: boolean;
  id_usergenerate: number;
  status_generate: boolean;
  blacklist: boolean;
  created_at?: string;
  updated_at?: string;
}